requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiostream==0.4.5
anyio==4.0.0
argilla==0.0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
chardet==5.2.0
chromadb==0.3.25
click==8.1.7
clickhouse-connect==0.5.25
cohere==4.7.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
ctransformers==0.2.5
Cython==3.0.0
dataclasses-json==0.5.7
debugpy==1.6.7.post1
decorator==5.1.1
Deprecated==1.2.14
docarray==0.21.0
docker==6.1.3
docstring-parser==0.15
duckdb==0.8.0
ecdsa==0.18.0
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.4
fake-useragent==1.1.3
fastapi==0.95.2
filelock==3.12.2
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.59.0
grpcio==1.47.5
grpcio-health-checking==1.47.5
grpcio-reflection==1.47.5
grpcio-tools==1.47.5
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hnswlib==0.7.0
hpack==4.0.0
httpcore==0.17.2
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==6.0.1
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jcloud==0.2.10
jedi==0.19.0
jina==3.15.2
jina-hubble-sdk==0.38.0
joblib==1.3.2
json5==0.9.14
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
langchain==0.0.186
langflow==0.0.81
llama-cpp-python==0.1.57
lxml==4.9.2
markdown-it-py==2.2.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.8.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.25.2
olefile==0.46
onnxruntime==1.15.0
openai==0.27.7
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-aiohttp-client==0.39b0
opentelemetry-instrumentation-asgi==0.39b0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation-grpc==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
overrides==7.3.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
portalocker==2.7.0
posthog==3.0.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pure-eval==0.2.2
pyarrow==11.0.0
pydantic==1.10.8
Pygments==2.15.1
pypandoc==1.11
pypdf==3.9.0
pyreadline3==3.4.1
pysrt==1.1.2
python-docx==0.8.11
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qdrant-client==1.2.0
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.8.8
reportlab==4.0.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.1
rpds-py==0.10.2
rsa==4.9
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.3.3
tinycss2==1.2.1
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
typer==0.7.0
types-PyYAML==6.0.12.10
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
unstructured==0.5.13
uri-template==1.3.0
uritemplate==4.1.1
uvicorn==0.20.0
validators==0.20.0
watchfiles==0.19.0
wcwidth==0.2.6
weaviate-client==3.19.2
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
widgetsnbextension==4.0.8
wikipedia==1.4.0
XlsxWriter==3.1.2
yarl==1.9.2
